requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.12.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
exceptiongroup==1.1.3
flake8==7.0.0
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
pathmatch==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
typing==3.7.4.3
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
